Drone debugging message changes: - No longer stores every single message ever received for a given   prog widget ID; not necessary, and could lead to out-of-memory   conditions if a drone is left idle for too long running the same   prog widget piece. - Now stores (and syncs to clients) the last message received for each   prog widget, timestamped by client when received - Drone debugger display shows time elapsed since last message received   for each prog widget. #197

+++ src/main/java/me/desht/pneumaticcraft/common/util/PneumaticCraftUtils.java
@@ -207,7 +207,7 @@ public class PneumaticCraftUtils {
      * @param fraction When true, 30 ticks will show as '1.5s' instead of '1s'.
      * @return a formatted time
      */
-    public static String convertTicksToMinutesAndSeconds(int ticks, boolean fraction) {
+    public static String convertTicksToMinutesAndSeconds(long ticks, boolean fraction) {
         String part = ticks % 20 * 5 + "";
         if (part.length() < 2) part = "0" + part;
         ticks /= 20;// first convert to seconds.
